requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
Pillow>=10.0.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
huggingface-hub>=0.16.0
pi-heif>=0.21.0
refiners>=0.4.0
celery>=5.3.0
redis>=5.0.0
django-cors-headers>=4.3.0
python-multipart>=0.0.6
drf-spectacular>=0.27.0